feat: Redesign action creators
The previous version of Redy has a fatal problem.
It uses `Function.prototype.name` as an action type
but it is often renamed to a short name (usually a single character)
in production environment by a source code minification tool, such as
UglifyJS. This causes several problems:

- Action type conflicts are more likey to occur and it is not uncontrollable.
- It makes debugging (in production) difficult because of mangled action types.

There are several options to avoid this. But none of them are desirable.

- Force users to disable function name mangling of JS minification.
    - Not practical and easy to forget.
- Use other value as an action type such as a function itself or the result of
  `Function.prototype.toString`.
    - It still makes debugging difficult.
    - It makes serialization of actions impossible or difficult.
      It violates Redux philosophy: https://redux.js.org/faq/actions#why-should-type-be-a-string-or-at-least-serializable-why-should-my-action-types-be-constants
- Provide a babel plugin which transforms action creator definitions to something like
  a definition which specifies an action type using a string literal.
    - Why we need Babel just for a state management library?
      Force users to use it? Awful.

So, we decide to re-design Redy to define an action type explicitly.
This makes code a bit verbose but simple and safe.